Examen2doP/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib[bcrypt]
python-multipart